Make some refraction of codebase
